+++ cluster_tools/cluster_tools/executor_protocol.py
@@ -31,6 +31,7 @@ class Executor(Protocol, AbstractContextManager["Executor"]):
         fn: Callable[[_S], _T],
         args: Iterable[_S],
         output_pickle_path_getter: Callable[[_S], PathLike] | None = None,
+        output_key_getter: Callable[[_S], str] | None = None,
     ) -> list[Future[_T]]: ...
 
     def map(